requirements.txt
docx-mailmerge==0.5.0
lxml==4.8.0
